moved map logic out
