+++ business_gen.py
@@ -10,14 +10,15 @@ import csv
 from pydantic import BaseModel
 from bs4 import BeautifulSoup
 
+from maps import geocode
+
 
 load_dotenv()
 GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")
 QUEUE_URL = os.getenv("QUEUE_URL")
-ENDPOINT = "https://maps.googleapis.com/maps/api/place/textsearch/json?"
-DETAILS_ENDPOINT = "https://maps.googleapis.com/maps/api/place/details/json?"
-GEOCODE_ENDPOINT = "https://maps.googleapis.com/maps/api/geocode/json?"
-NUM_DIVISIONS = 5  # Number of subdivisions in each dimension (change as needed)
+# ENDPOINT = "https://maps.googleapis.com/maps/api/place/textsearch/json?"
+# DETAILS_ENDPOINT = "https://maps.googleapis.com/maps/api/place/details/json?"
+# NUM_DIVISIONS = 5  # Number of subdivisions in each dimension (change as needed)
 requests.packages.urllib3.disable_warnings()
 
 
@@ -27,80 +28,84 @@ class BusinessData(BaseModel):
     email: str = ""
     web_content: str = ""
 
+# TODO: add type hints
 
-def geocode_location(location):
-    params = {"address": location, "key": GOOGLE_API_KEY}
-    response = requests.get(GEOCODE_ENDPOINT, params=params)
-    data = response.json()
-    if data["status"] == "OK":
-        lat = data["results"][0]["geometry"]["location"]["lat"]
-        lng = data["results"][0]["geometry"]["location"]["lng"]
-        return lat, lng
-    else:
-        raise ValueError(
-            f"Error geocoding location {location}. Error message: {data['status']}"
-        )
-
-
-def subdivide_region(min_lat, max_lat, min_lng, max_lng, num_divisions):
-    lat_step = (max_lat - min_lat) / num_divisions
-    lng_step = (max_lng - min_lng) / num_divisions
-    grids = []
-    for i in range(num_divisions):
-        for j in range(num_divisions):
-            grid_min_lat = min_lat + i * lat_step
-            grid_max_lat = grid_min_lat + lat_step
-            grid_min_lng = min_lng + j * lng_step
-            grid_max_lng = grid_min_lng + lng_step
-            grids.append((grid_min_lat, grid_max_lat, grid_min_lng, grid_max_lng))
-    return grids
+# def geocode_location(location):
+#     """geocode_location takes location [str] and calls Google Maps API to get 
+#     lat and long [int, int]
+#     """
+#     params = {"address": location, "key": GOOGLE_API_KEY}
+#     response = requests.get(GEOCODE_ENDPOINT, params=params)
+#     data = response.json()
+#     if data["status"] == "OK":
+#         lat = data["results"][0]["geometry"]["location"]["lat"]
+#         lng = data["results"][0]["geometry"]["location"]["lng"]
+#         return lat, lng
+#     else:
+#         raise ValueError(
+#             f"Error geocoding location {location}. Error message: {data['status']}"
+#         )
 
 
-def get_places_for_grid(query, min_lat, max_lat, min_lng, max_lng):
-    location = f"{(min_lat+max_lat)/2},{(min_lng+max_lng)/2}"  # Use the center of the grid as location
-    parameters = {
-        "query": query,
-        "location": location,
-        "key": GOOGLE_API_KEY,
-        "radius": 2000,
-    }  # Adjust the radius as needed
-    all_results = []
-    response = requests.get(ENDPOINT, params=parameters)
-    results = response.json().get("results", [])
-    all_results.extend(results)
-    # Handle pagination
-    while "next_page_token" in response.json():
-        parameters["pagetoken"] = response.json()["next_page_token"]
-        response = requests.get(ENDPOINT, params=parameters)
-        results = response.json().get("results", [])
-        all_results.extend(results)
-    return all_results
+# def subdivide_region(min_lat, max_lat, min_lng, max_lng, num_divisions):
+#     lat_step = (max_lat - min_lat) / num_divisions
+#     lng_step = (max_lng - min_lng) / num_divisions
+#     grids = []
+#     for i in range(num_divisions):
+#         for j in range(num_divisions):
+#             grid_min_lat = min_lat + i * lat_step
+#             grid_max_lat = grid_min_lat + lat_step
+#             grid_min_lng = min_lng + j * lng_step
+#             grid_max_lng = grid_min_lng + lng_step
+#             grids.append((grid_min_lat, grid_max_lat, grid_min_lng, grid_max_lng))
+#     return grids
 
 
-def get_places(query, min_lat, max_lat, min_lng, max_lng):
-    grids = subdivide_region(min_lat, max_lat, min_lng, max_lng, NUM_DIVISIONS)
-    all_results = []
-    for grid in grids:
-        grid_results = get_places_for_grid(query, *grid)
-        all_results.extend(grid_results)
-    # Removing potential duplicates
-    seen_place_ids = set()
-    unique_results = []
-    for place in all_results:
-        if place["place_id"] not in seen_place_ids:
-            seen_place_ids.add(place["place_id"])
-            unique_results.append(place)
-    # Fetch additional details for unique results
-    for idx, place in enumerate(unique_results):
-        details = get_place_details(place["place_id"])
-        unique_results[idx].update(details)
-    return unique_results
+# def get_places_for_grid(query, min_lat, max_lat, min_lng, max_lng):
+#     location = f"{(min_lat+max_lat)/2},{(min_lng+max_lng)/2}"  # Use the center of the grid as location
+#     parameters = {
+#         "query": query,
+#         "location": location,
+#         "key": GOOGLE_API_KEY,
+#         "radius": 2000,
+#     }  # Adjust the radius as needed
+#     all_results = []
+#     response = requests.get(ENDPOINT, params=parameters)
+#     results = response.json().get("results", [])
+#     all_results.extend(results)
+#     # Handle pagination
+#     while "next_page_token" in response.json():
+#         parameters["pagetoken"] = response.json()["next_page_token"]
+#         response = requests.get(ENDPOINT, params=parameters)
+#         results = response.json().get("results", [])
+#         all_results.extend(results)
+#     return all_results
 
 
-def get_place_details(place_id):
-    parameters = {"place_id": place_id, "key": GOOGLE_API_KEY}
-    response = requests.get(DETAILS_ENDPOINT, params=parameters)
-    return response.json().get("result", {})
+# def get_places(query, min_lat, max_lat, min_lng, max_lng):
+#     grids = subdivide_region(min_lat, max_lat, min_lng, max_lng, NUM_DIVISIONS)
+#     all_results = []
+#     for grid in grids:
+#         grid_results = get_places_for_grid(query, *grid)
+#         all_results.extend(grid_results)
+#     # Removing potential duplicates
+#     seen_place_ids = set()
+#     unique_results = []
+#     for place in all_results:
+#         if place["place_id"] not in seen_place_ids:
+#             seen_place_ids.add(place["place_id"])
+#             unique_results.append(place)
+#     # Fetch additional details for unique results
+#     for idx, place in enumerate(unique_results):
+#         details = get_place_details(place["place_id"])
+#         unique_results[idx].update(details)
+#     return unique_results
+
+
+# def get_place_details(place_id):
+#     parameters = {"place_id": place_id, "key": GOOGLE_API_KEY}
+#     response = requests.get(DETAILS_ENDPOINT, params=parameters)
+#     return response.json().get("result", {})
 
 
 def save_to_initial_queue(data, queue_url):
@@ -258,11 +263,11 @@ def save_to_sqs(data, queue_url):
 def main():
     query = input("Enter the type of business: ")
     location_name = input("Enter the city and state (e.g. 'Tacoma, WA'): ")
-    lat, lng = geocode_location(location_name)
+    lat, lng = geocode.geocode_location(location_name)
     delta = 0.05  # Adjust this value as needed for city size
     min_lat, max_lat = lat - delta, lat + delta
     min_lng, max_lng = lng - delta, lng + delta
-    data = get_places(query, min_lat, max_lat, min_lng, max_lng)
+    data = geocode.get_places(query, min_lat, max_lat, min_lng, max_lng)
     save_to_initial_queue(data, QUEUE_URL)
     save_to_sqs(data, QUEUE_URL)
     # added csv to prove we are getting emails
